refactor: member service 통합
