+++ src/main/java/com/planetrush/planetrush/member/service/OAuthService.java
@@ -1,8 +1,9 @@
 package com.planetrush.planetrush.member.service;
 
 import com.planetrush.planetrush.member.service.dto.LoginDto;
+import com.planetrush.planetrush.member.service.dto.ReissueDto;
 
-public interface AuthService {
+public interface OAuthService {
 
 	/**
 	 * 카카오 로그인을 진행합니다.
@@ -24,4 +25,11 @@ public interface AuthService {
 	 * @param memberId 유저의 고유 id
 	 */
 	void withdrawnMember(Long memberId);
+
+	/**
+	 * 주어진 리프레시 토큰을 사용하여 새로운 액세스 토큰과 리프레시 토큰을 발급합니다.
+	 * @param refreshToken 현재 사용자의 리프레시 토큰
+	 * @return 새로운 액세스 토큰과 리프레시 토큰을 담고 있는 ReissueDto 객체
+	 */
+	ReissueDto reissueToken(String refreshToken);
 }
